+++ R/data.R
@@ -6,7 +6,7 @@
 #'
 #' @format A data frame with `r nrow(elements::ALEData)` rows and `r ncol(elements::ALEData)` columns, the definitions of which are:
 #' \describe{
-#'   \item{taxon_code}{The taxon, see `elements::ModelTaxa`.}
+#'   \item{taxon_code}{The taxon, see `elements::TaxonomicBackbone`.}
 #'   \item{x}{The variable value.}
 #'   \item{y}{The ALE value.}
 #'   \item{variable}{The variable name.}
@@ -15,7 +15,7 @@
 
 #' EuroSL data
 #' 
-#' Selected columns for the accepted taxa in the European European Species List (EuroSL) database.
+#' Selected columns for the accepted taxa in the European Species List (EuroSL) database.
 #'
 #' \code{EuroSL} 
 #'
@@ -50,7 +50,7 @@
 #'   \item{bio06}{Minimum temperature in the coldest month}
 #'   \item{bio16}{Precipitation in the wettest quarter}
 #'   \item{bio17}{Precipitation in the driest quarter}
-#'   \item{taxon_code}{The taxon, see `elements::ModelTaxa`.}
+#'   \item{taxon_code}{The taxon, see `elements::TaxonomicBackbone`.}
 #' }
 #' @references{
 #' Copernicus Climate Change Service, 2021. Downscaled bioclimatic indicators for selected regions from 1950 to 2100 derived from climate projections. https://doi.org/10.24381/CDS.0AB27596
@@ -83,7 +83,7 @@
 #'   \item{bio06}{Minimum temperature in the coldest month}
 #'   \item{bio16}{Precipitation in the wettest quarter}
 #'   \item{bio17}{Precipitation in the driest quarter}
-#'   \item{taxon_code}{The taxon, see `elements::ModelTaxa`.}
+#'   \item{taxon_code}{The taxon, see `elements::TaxonomicBackbone`.}
 #' }
 #' @references{
 #' Copernicus Climate Change Service, 2021. Downscaled bioclimatic indicators for selected regions from 1950 to 2100 derived from climate projections. https://doi.org/10.24381/CDS.0AB27596
@@ -95,25 +95,6 @@
 #' Tichý, L., Axmanová, I., Dengler, J., Guarino, R., Jansen, F., Midolo, G., Nobis, M.P., Van Meerbeek, K., Aćić, S., Attorre, F., Bergmeier, E., Biurrun, I., Bonari, G., Bruelheide, H., Campos, J.A., Čarni, A., Chiarucci, A., Ćuk, M., Ćušterevska, R., Didukh, Y., Dítě, D., Dítě, Z., Dziuba, T., Fanelli, G., Fernández-Pascual, E., Garbolino, E., Gavilán, R.G., Gégout, J.-C., Graf, U., Güler, B., Hájek, M., Hennekens, S.M., Jandt, U., Jašková, A., Jiménez-Alfaro, B., Julve, P., Kambach, S., Karger, D.N., Karrer, G., Kavgacı, A., Knollová, I., Kuzemko, A., Küzmič, F., Landucci, F., Lengyel, A., Lenoir, J., Marcenò, C., Moeslund, J.E., Novák, P., Pérez-Haase, A., Peterka, T., Pielech, R., Pignatti, A., Rašomavičius, V., Rūsiņa, S., Saatkamp, A., Šilc, U., Škvorc, Ž., Theurillat, J.-P., Wohlgemuth, T., Chytrý, M., 2023. Ellenberg-type indicator values for European vascular plant species. Journal of Vegetation Science 34, e13168. https://doi.org/10.1111/jvs.13168
 #' }
 "ExampleData2"
-
-#' Information on the modelled taxa.
-#' 
-#' Information on the modelled taxa in `elements`, including the taxon_code values underpinning all functions and objects.
-#' 
-#' Note that further information will be included before v1 of the package is minted.
-#'
-#' \code{ModelTaxa} 
-#'
-#' @format A data frame with `r nrow(elements::ModelTaxa)` rows and `r ncol(elements::ModelTaxa)` columns, the definitions of which are:
-#' \describe{
-#'   \item{taxon_name}{The full taxon name, currently as present in the "Matched concept" column in the EVA.}
-#'   \item{taxon_code}{The taxon code, equivalent to the full taxon name but all lower case and with whitespace replaced with underscores.}
-#'   \item{eurosl_name}{The taxon name as it appears in EuroSL, where there is a match.}
-#'   \item{eurosl_status}{The taxonomic status of the eurosl_name.}
-#'   \item{eurosl_id}{The ID of the eurosl_name.}
-#'   \item{n_presences_eva}{The number of presences in suitable samples from the EVA.}
-#' }
-"ModelTaxa"
 
 #' Niche width data
 #'
@@ -136,7 +117,7 @@
 #'   \item{q90}{The 90% quantile value.}
 #'   \item{q95}{The 95% quantile value.}
 #'   \item{q99}{The 99% quantile value.}
-#'   \item{taxon_code}{The taxon, see `elements::ModelTaxa`.}
+#'   \item{taxon_code}{The taxon, see `elements::TaxonomicBackbone`.}
 #' }
 "NicheWidths"
 
@@ -148,7 +129,7 @@
 #'
 #' @format A data frame with `r nrow(elements::PerformanceMeasures)` rows and `r ncol(elements::PerformanceMeasures)` columns, the definitions of which are:
 #' \describe{
-#'   \item{taxon_code}{The taxon, see `elements::ModelTaxa`.}
+#'   \item{taxon_code}{The taxon, see `elements::TaxonomicBackbone`.}
 #'   \item{Holdout.PrecisionRecallAreaUnderCurve}{The Precision-Recall Area Under the Curve (PRAUC), calculated using the random holdout sample test data.}
 #'   \item{Holdout.Precision}{The Precision, calculated using the random holdout sample test data.}
 #'   \item{Holdout.Recall}{The Recall, calculated using the random holdout sample test data.}
@@ -167,12 +148,48 @@
 #'
 #' @format A data frame with `r nrow(elements::PDPData)` rows and `r ncol(elements::PDPData)` columns, the definitions of which are:
 #' \describe{
-#'   \item{taxon_code}{The taxon, see `elements::ModelTaxa`.}
+#'   \item{taxon_code}{The taxon, see `elements::TaxonomicBackbone`.}
 #'   \item{x}{The variable value.}
 #'   \item{y}{The PDP value.}
 #'   \item{variable}{The variable name.}
 #' }
 "PDPData"
+
+#' Taxonomic backbone
+#' 
+#' The taxon names and codes for the modeled taxa in the EVA, with the associated taxon concept information from EuroSL, and parent taxa.
+#'
+#' \code{TaxonomicBackbone} 
+#'
+#' @format A data frame with `r nrow(elements::TaxonomicBackbone)` rows and `r ncol(elements::TaxonomicBackbone)` columns, the definitions of which are:
+#' \describe{
+#'   \item{taxon_code}{The taxon codes used throughout the package, formed from the taxon_name values by coercing all letters to lower and replacing whitespace with underscores.}
+#'   \item{taxon_name}{The taxon names for the modelled taxa, as they appear in the EVA.}
+#'   \item{taxon_concept}{The acccepted taxon concepts for each taxon_name value, following the authority of the EuroSL - see `elements::EuroSL`}
+#'   \item{eurosl_status}{The taxonomic status of the taxon_name, following the authority of the EuroSL.}
+#'   \item{eurosl_id}{The EuroSL ID code for the taxon_name.}
+#'   \item{Species}{The Species name associated with the taxon_name.}
+#'   \item{Genus}{The parent Genus taxon associated with the taxon_name.}
+#'   \item{Tribe}{The parent Tribe taxon associated with the taxon_name.}
+#'   \item{Family}{The parent Family taxon associated with the taxon_name.}
+#'   \item{Order}{The parent Order taxon associated with the taxon_name.}
+#'   \item{Superorder}{The parent Superorder taxon associated with the taxon_name.}
+#'   \item{Class}{The parent Class taxon associated with the taxon_name.}
+#'   \item{Subdivision}{The parent Subdivision taxon associated with the taxon_name.}
+#'   \item{Phylum}{The parent Phylum taxon associated with the taxon_name.}
+#'   \item{Root}{The parent Root taxon associated with the taxon_name.}
+#'   \item{Section}{The Section name associated with the taxon_name.}
+#'   \item{Species Aggregate}{The Species Aggregate name associated with the taxon_name.}
+#'   \item{Subclass}{The parent Subclass taxon associated with the taxon_name.}
+#'   \item{Division}{The parent Division taxon associated with the taxon_name.}
+#'   \item{Suprageneric Taxon}{The parent Suprageneric Taxon taxon associated with the taxon_name.}
+#'   \item{Subfamily}{The parent Subfamily taxon associated with the taxon_name.}
+#'   \item{Subspecies}{The Subspecies name associated with the taxon_name.}
+#'   \item{Variety}{The Variety name associated with the taxon_name.}
+#'   \item{Coll. species}{The Coll. species name associated with the taxon_name.}
+#'   \item{n_presences_eva}{The number of presences of the taxon (taxon_name) in the EVA, only considering plots with valid coordinates and years.}
+#' }
+"TaxonomicBackbone"
 
 #' Environmental indicator value data
 #'
@@ -195,13 +212,13 @@
 #' }
 "VariableData"
 
-#' Environmental indicator value names
+#' Environmental variable names and codes
 #'
-#' A dataset containing a lookup for the Ecological Indicator Value (EIV) codes used throughout the `elements` package and full variable namesnames.
+#' A dataset containing a lookup for the Ecological Indicator Value (EIV) and bioclimatic variable codes used throughout the `elements` package and full variable namesnames.
 #'
-#' \code{VariableNames} 
+#' \code{VariableLookup} 
 #'
-#' @format A data frame with `r nrow(elements::VariableNames)` rows and `r ncol(elements::VariableNames)` columns, the definitions of which are:
+#' @format A data frame with `r nrow(elements::VariableLookup)` rows and `r ncol(elements::VariableLookup)` columns, the definitions of which are:
 #' \describe{
 #'   \item{raw_name}{The raw name of the variable as present in the parent dataset.}
 #'   \item{variable_code}{The variable code, used throughout the `elements` R package.}
@@ -209,4 +226,14 @@
 #'   \item{variable_plot_name}{A 'tidy' variable name with spaces.}
 #'   \item{model_var}{A boolean indicating whether the variable was included as a predictor.}
 #' }
+"VariableLookup"
+
+#' Model variables codes
+#'
+#' A vector containing the environmental variable codes for the model variables; 
+#' equivalent to the variable_code values in the `elements::VariableLookup` data frame for rows where model_var is TRUE.
+#'
+#' \code{VariableNames} 
+#'
+#' @format A vector containing `r nrow(elements::VariableNames)` strings.
 "VariableNames"